requirements.txt
streamlit
pandas
xlrd
numpy
openpyxl
pydeck
plotly.express
